change vtx-idx to int64

+++ src/hyteg/adaptiverefinement/simplexFactory.cpp
@@ -13,7 +13,7 @@ std::shared_ptr< Simplex1 > SimplexFactory::make_edge( int a, int b, Color color
 
 std::shared_ptr< Simplex2 > SimplexFactory::make_face( int a, int b, int c )
 {
-   std::array< int, 3 >                         V{ _vertices[a], _vertices[b], _vertices[c] };
+   std::array< int64_t, 3 >                     V{ _vertices[a], _vertices[b], _vertices[c] };
    std::array< std::shared_ptr< Simplex1 >, 3 > E{ _edges[{ a, b }], _edges[{ b, c }], _edges[{ c, a }] };
 
    auto face = std::make_shared< Simplex2 >( V, E );
@@ -23,7 +23,7 @@ std::shared_ptr< Simplex2 > SimplexFactory::make_face( int a, int b, int c )
 
 std::shared_ptr< Simplex3 > SimplexFactory::make_cell( int a, int b, int c, int d )
 {
-   std::array< int, 4 > V{ _vertices[a], _vertices[b], _vertices[c], _vertices[d] };
+   std::array< int64_t, 4 > V{ _vertices[a], _vertices[b], _vertices[c], _vertices[d] };
 
    std::array< std::shared_ptr< Simplex1 >, 6 > E{
        _edges[{ a, b }], _edges[{ b, c }], _edges[{ c, a }], _edges[{ a, d }], _edges[{ b, d }], _edges[{ c, d }] };
